Fix #4121: particle_file_for_external_lattice in controls
