+++ sirepo/pkcli/controls.py
@@ -8,6 +8,7 @@ from __future__ import absolute_import, division, print_function
 from pykern import pkio, pkjson
 from pykern.pkcollections import PKDict
 from sirepo.template import template_common
+import sirepo.pkcli.madx
 import sirepo.template.controls as template
 from sirepo import simulation_db
 
@@ -28,3 +29,19 @@ def run(cfg_dir):
 
 def run_background(cfg_dir):
     template_common.exec_parameters()
+
+
+def particle_file_for_external_lattice():
+    data = simulation_db.read_json(
+        template_common.INPUT_BASE_NAME,
+    )
+    b = [k for k in data.models.command_twiss
+        if k in data.models.externalLattice.models.bunch]
+    for p in b:
+        data.models.externalLattice.models.bunch[p] = data.models.command_twiss[p]
+    beam = data.models.command_beam
+    data = data.models.externalLattice
+    data.models.command_beam = beam
+    data.report = 'unused'
+    data.models.bunch.matchTwissParameters = '0'
+    sirepo.pkcli.madx.create_particle_file(pkio.py_path('.'), data)
